requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
python-dotenv==1.2.1
google-genai==1.52.0
google-auth==2.43.0
google-cloud-aiplatform==1.128.0
Pillow==12.0.0
gunicorn==23.0.0

